fixed a casting error from pointer to int

+++ cpu/trap.c
@@ -14,29 +14,29 @@ static void init_idt_entry(struct IdtEntry *entry, uint64_t addr, uint8_t attrib
 
 void init_idt(void)
 {
-    init_idt_entry(&vectors[0],(uint64_t)vector0,0x8e);
-    init_idt_entry(&vectors[1],(uint64_t)vector1,0x8e);
-    init_idt_entry(&vectors[2],(uint64_t)vector2,0x8e);
-    init_idt_entry(&vectors[3],(uint64_t)vector3,0x8e);
-    init_idt_entry(&vectors[4],(uint64_t)vector4,0x8e);
-    init_idt_entry(&vectors[5],(uint64_t)vector5,0x8e);
-    init_idt_entry(&vectors[6],(uint64_t)vector6,0x8e);
-    init_idt_entry(&vectors[7],(uint64_t)vector7,0x8e);
-    init_idt_entry(&vectors[8],(uint64_t)vector8,0x8e);
-    init_idt_entry(&vectors[10],(uint64_t)vector10,0x8e);
-    init_idt_entry(&vectors[11],(uint64_t)vector11,0x8e);
-    init_idt_entry(&vectors[12],(uint64_t)vector12,0x8e);
-    init_idt_entry(&vectors[13],(uint64_t)vector13,0x8e);
-    init_idt_entry(&vectors[14],(uint64_t)vector14,0x8e);
-    init_idt_entry(&vectors[16],(uint64_t)vector16,0x8e);
-    init_idt_entry(&vectors[17],(uint64_t)vector17,0x8e);
-    init_idt_entry(&vectors[18],(uint64_t)vector18,0x8e);
-    init_idt_entry(&vectors[19],(uint64_t)vector19,0x8e);
-    init_idt_entry(&vectors[32],(uint64_t)vector32,0x8e);
-    init_idt_entry(&vectors[39],(uint64_t)vector39,0x8e);
+    init_idt_entry(&vectors[0],(intptr_t)vector0,0x8e);
+    init_idt_entry(&vectors[1],(intptr_t)vector1,0x8e);
+    init_idt_entry(&vectors[2],(intptr_t)vector2,0x8e);
+    init_idt_entry(&vectors[3],(intptr_t)vector3,0x8e);
+    init_idt_entry(&vectors[4],(intptr_t)vector4,0x8e);
+    init_idt_entry(&vectors[5],(intptr_t)vector5,0x8e);
+    init_idt_entry(&vectors[6],(intptr_t)vector6,0x8e);
+    init_idt_entry(&vectors[7],(intptr_t)vector7,0x8e);
+    init_idt_entry(&vectors[8],(intptr_t)vector8,0x8e);
+    init_idt_entry(&vectors[10],(intptr_t)vector10,0x8e);
+    init_idt_entry(&vectors[11],(intptr_t)vector11,0x8e);
+    init_idt_entry(&vectors[12],(intptr_t)vector12,0x8e);
+    init_idt_entry(&vectors[13],(intptr_t)vector13,0x8e);
+    init_idt_entry(&vectors[14],(intptr_t)vector14,0x8e);
+    init_idt_entry(&vectors[16],(intptr_t)vector16,0x8e);
+    init_idt_entry(&vectors[17],(intptr_t)vector17,0x8e);
+    init_idt_entry(&vectors[18],(intptr_t)vector18,0x8e);
+    init_idt_entry(&vectors[19],(intptr_t)vector19,0x8e);
+    init_idt_entry(&vectors[32],(intptr_t)vector32,0x8e);
+    init_idt_entry(&vectors[39],(intptr_t)vector39,0x8e);
 
     idt_pointer.limit = sizeof(vectors)-1;
-    idt_pointer.addr = (uint64_t)vectors;
+    idt_pointer.addr = (intptr_t)vectors;
     load_idt(&idt_pointer);
 }
